+++ Assets/PurrUI/CHANGELOG.md
@@ -1,3 +1,66 @@
+# [1.3.0-beta.8](https://github.com/PurrNet/PurrUI/compare/v1.3.0-beta.7...v1.3.0-beta.8) (2026-03-12)
+
+
+### Bug Fixes
+
+* add component and create menu item ([dab3c79](https://github.com/PurrNet/PurrUI/commit/dab3c79121a2e75e0a8907cd2e525be7ab5b3ec1))
+
+# [1.3.0-beta.7](https://github.com/PurrNet/PurrUI/compare/v1.3.0-beta.6...v1.3.0-beta.7) (2026-03-11)
+
+
+### Bug Fixes
+
+* ensure additional canvas channels are initialized in SignedDistanceFieldGraphic ([205e5b6](https://github.com/PurrNet/PurrUI/commit/205e5b6a20ebbd2f1c320515706c879179f09111))
+
+# [1.3.0-beta.6](https://github.com/PurrNet/PurrUI/compare/v1.3.0-beta.5...v1.3.0-beta.6) (2026-03-10)
+
+
+### Bug Fixes
+
+* correct order of roundness fields in Glow and Rectangle graphic editors ([ca20640](https://github.com/PurrNet/PurrUI/commit/ca206408f11015eea6b80733ab223c37cbd8e8e3))
+
+# [1.3.0-beta.5](https://github.com/PurrNet/PurrUI/compare/v1.3.0-beta.4...v1.3.0-beta.5) (2026-03-09)
+
+
+### Features
+
+* refactor view stack management with Replace and AddToStack methods ([fd06b8f](https://github.com/PurrNet/PurrUI/commit/fd06b8ff6541682ae08f2dbd729264927a6a522c))
+
+# [1.3.0-beta.4](https://github.com/PurrNet/PurrUI/compare/v1.3.0-beta.3...v1.3.0-beta.4) (2026-03-09)
+
+
+### Features
+
+* add OnPushed and OnPopped methods with transition handling for MonoView ([f7f1012](https://github.com/PurrNet/PurrUI/commit/f7f1012ce3ed107ed652f3fdf5beaef6d9d20f36))
+
+# [1.3.0-beta.3](https://github.com/PurrNet/PurrUI/compare/v1.3.0-beta.2...v1.3.0-beta.3) (2026-03-09)
+
+
+### Features
+
+* add OnPushed and OnPopped methods for view stack management ([6181569](https://github.com/PurrNet/PurrUI/commit/6181569f622ce8cb3537399e8dcc6e1f588f8280))
+
+# [1.3.0-beta.2](https://github.com/PurrNet/PurrUI/compare/v1.3.0-beta.1...v1.3.0-beta.2) (2026-03-09)
+
+
+### Bug Fixes
+
+* clear selected GameObject when moving MonoView to background ([b154b3c](https://github.com/PurrNet/PurrUI/commit/b154b3cf6cce3552a4f8a4ad9fff5ecb7ae4011d))
+* update MoveToBackground and MoveToForeground methods for improved interaction handling ([1644bd1](https://github.com/PurrNet/PurrUI/commit/1644bd160bdd831d35fdb86caa28f9bfe774e619))
+* ViewTransitions, just assume midle of screen ([e83da40](https://github.com/PurrNet/PurrUI/commit/e83da40cc9db22bdc1019f49d0d78a0425111271))
+
+
+### Features
+
+* add culling and unculling transitions for MonoView management ([604b51d](https://github.com/PurrNet/PurrUI/commit/604b51df071c6f6e7db68682a6a5f697e251ca25))
+
+# [1.3.0-beta.1](https://github.com/PurrNet/PurrUI/compare/v1.2.0...v1.3.0-beta.1) (2026-03-08)
+
+
+### Features
+
+* add isTopMost property and top method to MonoView and ViewStack for improved stack management ([add8d8f](https://github.com/PurrNet/PurrUI/commit/add8d8f99cbe85c0bac188f70df9290ef86b18dd))
+
 # [1.2.0](https://github.com/PurrNet/PurrUI/compare/v1.1.0...v1.2.0) (2026-03-08)